requirements.txt
asgiref==3.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.17
django-bootstrap3==15.0.0
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.7.0
django-utils-six==2.0
djangorestframework==3.11.2
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
